// src/api/orval/model/serverCertificate.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { CertificateStatus } from "./certificateStatus";

export interface ServerCertificate {
  vpnServerId?: number;
  /** @nullable */
  commonName?: string | null;
  status?: CertificateStatus;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  unknownField?: string | null;
  isRevoked?: boolean;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  certificatePath?: string | null;
  /** @nullable */
  keyPath?: string | null;
  expiryDate?: string;
  /** @nullable */
  revokeDate?: string | null;
}